+++ packages/next/src/utils.ts
@@ -2110,14 +2110,14 @@ export const onPrerenderRoute =
 
     // If enabled, try to get the postponed route information from the file
     // system and use it to assemble the prerender.
-    let prerender: string | undefined;
+    let postponedPrerender: string | undefined;
     if (experimentalPPR && appDir) {
       const htmlPath = path.join(appDir, `${routeFileNoExt}.html`);
       const metaPath = path.join(appDir, `${routeFileNoExt}.meta`);
       if (fs.existsSync(htmlPath) && fs.existsSync(metaPath)) {
         const meta = JSON.parse(await fs.readFile(metaPath, 'utf8'));
         if ('postponed' in meta && typeof meta.postponed === 'string') {
-          prerender = meta.postponed;
+          postponedPrerender = meta.postponed;
 
           // Assign the headers Content-Type header to the prerendered type.
           initialHeaders ??= {};
@@ -2127,7 +2127,7 @@ export const onPrerenderRoute =
 
           // Read the HTML file and append it to the prerendered content.
           const html = await fs.readFileSync(htmlPath, 'utf8');
-          prerender += html;
+          postponedPrerender += html;
         }
       }
 
@@ -2144,14 +2144,14 @@ export const onPrerenderRoute =
       }
     }
 
-    if (prerender) {
+    if (postponedPrerender) {
       const contentType = initialHeaders?.['content-type'];
       if (!contentType) {
         throw new Error("Invariant: contentType can't be undefined");
       }
 
       // Assemble the prerendered file.
-      htmlFsRef = new FileBlob({ contentType, data: prerender });
+      htmlFsRef = new FileBlob({ contentType, data: postponedPrerender });
     } else if (
       appDir &&
       !dataRoute &&
@@ -2215,7 +2215,14 @@ export const onPrerenderRoute =
                     ? addLocaleOrDefault('/404.html', routesManifest, locale)
                     : '/404.html'
                   : isAppPathRoute
-                  ? prefetchDataRoute || dataRoute
+                  ? // When experimental PPR is enabled, we expect that the data
+                    // that should be served as a part of the prerender should
+                    // be from the prefetch data route. If this isn't enabled
+                    // for ppr, the only way to get the data is from the data
+                    // route.
+                    experimentalPPR
+                    ? prefetchDataRoute
+                    : dataRoute
                   : routeFileNoExt + '.json'
               }`
             ),
@@ -2272,10 +2279,6 @@ export const onPrerenderRoute =
       throw new Error('Invariant: expected to find prefetch data route PPR');
     }
 
-    // When the prefetch data path is available, use it for the prerender,
-    // otherwise use the data path.
-    const outputPrerenderPathData = outputPathPrefetchData || outputPathData;
-
     if (isSharedLambdas) {
       const outputSrcPathPage = normalizeIndexOutput(
         path.join(
@@ -2328,8 +2331,14 @@ export const onPrerenderRoute =
         htmlFsRef.contentType = htmlContentType;
         prerenders[outputPathPage] = htmlFsRef;
 
-        if (outputPrerenderPathData) {
-          prerenders[outputPrerenderPathData] = jsonFsRef;
+        if (outputPathPrefetchData) {
+          prerenders[outputPathPrefetchData] = jsonFsRef;
+        }
+
+        // If experimental ppr is not enabled for this route, then add the data
+        // route as a target for the prerender as well.
+        if (outputPathData && !experimentalPPR) {
+          prerenders[outputPathData] = jsonFsRef;
         }
       }
     }
@@ -2465,21 +2474,20 @@ export const onPrerenderRoute =
           : {}),
       });
 
-      if (outputPrerenderPathData) {
-        let normalizedPathData = outputPrerenderPathData;
-
+      const normalizePathData = (pathData: string) => {
         if (
           (srcRoute === '/' || srcRoute == '/index') &&
-          outputPrerenderPathData.endsWith(RSC_PREFETCH_SUFFIX)
+          pathData.endsWith(RSC_PREFETCH_SUFFIX)
         ) {
-          delete lambdas[normalizedPathData];
-          normalizedPathData = normalizedPathData.replace(
-            /([^/]+\.prefetch\.rsc)$/,
-            '__$1'
-          );
+          delete lambdas[pathData];
+          return pathData.replace(/([^/]+\.prefetch\.rsc)$/, '__$1');
         }
 
-        prerenders[normalizedPathData] = new Prerender({
+        return pathData;
+      };
+
+      if (outputPathData || outputPathPrefetchData) {
+        const prerender = new Prerender({
           expiration: initialRevalidate,
           lambda,
           allowQuery,
@@ -2500,21 +2508,30 @@ export const onPrerenderRoute =
                   ...initialHeaders,
                   'content-type': rscContentTypeHeader,
                   vary: rscVaryHeader,
-                  // If it contains a pre-render, then it was postponed.
-                  ...(prerender && rscDidPostponeHeader
+                  ...(postponedPrerender && rscDidPostponeHeader
                     ? { [rscDidPostponeHeader]: '1' }
                     : {}),
                 },
               }
             : {}),
         });
+
+        if (outputPathPrefetchData) {
+          prerenders[normalizePathData(outputPathPrefetchData)] = prerender;
+        }
+
+        // If experimental ppr is not enabled for this route, then add the data
+        // route as a target for the prerender as well.
+        if (outputPathData && !experimentalPPR) {
+          prerenders[normalizePathData(outputPathData)] = prerender;
+        }
       }
 
       // we need to ensure all prerenders have a matching .rsc output
       // otherwise routing could fall through unexpectedly for the
       // fallback: false case as it doesn't have a dynamic route
       // to catch the `.rsc` request for app -> pages routing
-      if (outputPrerenderPathData?.endsWith('.json') && appDir) {
+      if (outputPathData?.endsWith('.json') && appDir) {
         const dummyOutput = new FileBlob({
           data: '{}',
           contentType: 'application/json',